requirements.txt
python-telegram-bot==13.13
requests==2.28.1
beautifulsoup4==4.11.1
feedparser
sqlalchemy
sqlalchemy_serializer
